*: removed xasync_result *: added test for xfileinfo *: xfiledata that becomes free is now pushed on the free queue by the io thread, this makes the update and sync functions obsolete

+++ source/test/cpp/test_xfiledevice.cpp
@@ -498,7 +498,7 @@ namespace xcore
 					testFile->mFileLength = testFile->mMaxFileLength;
 				}
 
-				x_memcopy(testFile->mFileData + pos, buffer, bytesToWrite);
+				x_memcopy(testFile->mFileData + pos, buffer, (s32)bytesToWrite);
 				outNumBytesWritten = bytesToWrite;
 				return true;
 			}
